Create Windows.Carving.Qbot
This artifact yara-scans memory or unpacked DLL samples for Qbot (or Qakbot) detections, decrypts and returns the botnet and C2 configurations.
    
You may select specific file paths or processes to be yara-scanned, or allow it to yara-scan all memory space.
    
Currently, this artifact parses encrypted configurations from the Qbot variant introduced in April 2021. It will not parse the configurations from earlier variants.