Create Windows.Carving.IcedID.yaml (#230)
This artifact yara-scans memory or unpacked DLL samples for IcedID trojan (also known as BokBot, or Anubis) detections, finds, decodes and returns the Campaign ID and C2 configurations.

You may select specific file paths or processes to be yara-scanned, or allow it to yara-scan all memory space.

NOTE: This content simply carves the configuration and does not unpack files on disk. That means pointing this artifact as a packed or obfuscated file will not obtain the expected results.