Update artifact for Chainsaw v.1.1.7 (#284)
In my test case of this artifact, Chainsaw v.1.0.2 was returning very few detections (8) on a set of EVTX full of threats whereas v.1.1.7 returned 40+ detections. However, v.1.1.7 changed the behavior of the `--json` option which broke this original artifact. This VQL update replaces Chainsaw v.1.0.2 with v.1.1.7 and updates the CLI options. 

**WARNING!** As of the latest version of Chainsaw, there is a [bug](https://github.com/countercept/chainsaw/pull/62) that **does not return all detections via JSON output** which means **this VQL artifact is likely not to return complete detection results** until the bug is fixed! I have added this warning to the VQL description as it's very important for the analyst to know when using this artifact. Once the bug is confirmed fixed, we can update the download version and remove the warning. See [screenshot](https://twitter.com/eric_capuano/status/1527943034045874177?s=20&t=Ql8QSSH5slyJGBMtR2UQww).