Update PrintNightmare.yaml
According to Microsoft, this vulnerability can only be exploited if the “NoWarningNoElevationOnInstall” key in the registry is set to 1. 
   
The artifact scans the device registry to check if the beforementioned key exists or not; if it is undefined or doesn’t exist, then the system is not vulnerable to the PrintNightmare. Otherwise, the system is considered to be vulnerable to exploitation.

This vulnerability can be exploited using the Evil Printer attack.

The suggested VQL query looks for the registry values to find a registry key named “NoWarningNoElevationOnInstall”.
   
References:
   
https://msrc.microsoft.com/update-guide/vulnerability/CVE-2021-34527