Add artifact parsing GPG keys (#649)
PGP keys are used extensively in Debian package management. Sometimes
it's useful to inspect these keys. This artifact parses the output of
the "gpg" binary when operated on .gpg public key files in typical
keyring locations on Debian-based systems.

gpg is not treated as a tool because it will always exist on systems
where I imagine this artifact being run. Tell me if adding it as a tool
would still be the better approach.


![2023-08-08_15-29](https://github.com/Velocidex/velociraptor-docs/assets/4006538/e7b67a8e-aa79-4708-a0af-709970affbe9)