Force exchange zip to zero timestamps
We build the exchange zip file on each commit and previously encoded
timestamps in the zip - this means each release had a different zip
regardless of if the content of the exchange actually changed.

This causes a lot of churn in the repo and also makes it hard to know
when the exchange actually changed since the bundle is different on
each push.