Create Windows.Applications.Edge.History2023.yaml
Good morning, I have detected that the plugin to review Edge's history was not working correctly, as it was not displaying all the fields. I have modified the query to display all the consulted URLs by selecting the URL field within the history file. In my case, I use it for my own phishing process, and this makes it much easier.