Update PowershellMonitoring.yaml ScriptIgnorePath behavior
While helping a team tune out false positives from authorized scripts, we noticed that using the IgnoreScriptPath arg would not work for certain events, because the path to exclude was in the ScriptName field.  So far I've only seen this happen with EID 7937. There's no Path field, and CommandPath is present but empty.