Update 'Server.Utills.BackupAzure' to upload to subdirectory (#759)
Currently all data is uploaded to the Azure Blob Storage in the same
directory. However, if you run multiple hunts and let all of them upload
this becomes a big mess. This PR allows you to upload everything into a
sub-directory per flow, bundling the results of all hosts from the same
hunt/flow in the same subdirectory.