added artifact for parsing FreeBSD utx files (#903)
Hi, 

I wrote an artifact to parse the utx file under FreeBSD. 
The file is similar to the wtmp file under Linux. It records logins,
logouts, boots, shutdowns and system time changes.

I appended the results from running the artifact on a test system
(freebsd 14.1-RELEASE-p3 amd64)

[example_results.json](https://github.com/user-attachments/files/16818996/example_results.json)

Im not sure, if this artifact is better of in the Artifact Exchange or
shipped by default with Velociraptor. For now, I offer it here.
If there is potential to improve the VQL, please comment so :)

Kind Regards