Add API diagram (#947)
The diagram itself is not very important. The point of this is really to
test the SVG fonts. It contains all 3 fonts so it should be a good test
case.

- This diagram was created in my local Excalidraw instance
- then upgraded to be font-compatible with excalidraw.com using my
`Notebooks.UpdateExcalidrawFontMapping` notebook
- then opened and exported to SVG with excalidraw.com so that it
contains their font names (which refer to the font definitions that we
swap out)
- then using my `Notebooks.UpdateSVGfonts` notebook I swapped out the
font defs, although that will make no difference because Hugo will swap
them a 2nd time.