Update Windows.EventLogs.WonkaVision artifact
Artifact had an option to call another artifact named `Artifact.Custom.Windows.Kerberos.WonkaVision` but that artifact does not exist. It was probably the old name of `Windows.Detection.WonkaVision` and I updated the artifact accordingly. This artifact also tried to use `Windows.EventLogs.EvtxHunter` artifact with outdated parameters so I also updated parameters and usage of parameters in the artifact.