Refactor Artifacts, VQL and Notebooks sections (#990)
The goal of this refactor is to allow these core topics to be expanded
more easily in future.

Artifacts, VQL and Notebooks are now much more complex than they used to
be so they can't all be squashed together under the "VQL" topic. As core
concepts they deserve their own sections. This also allows us to expand
these topics without creating another level in the hierarchy.

I've created stub pages in the new sections for proposed future topics
and marked them as draft (so they are hidden for now). Stubs for Sigma
and YARA are also added under the VQL section, which is where I think
they should be located.


![refactor20250125](https://github.com/user-attachments/assets/f10a1c45-9ffe-4e42-9a85-8bcd0d59bd8e)

The new sections also mirror the related reference pages. This allows
readers to see and access the relevant references while they are
engaging with documentation on a topic.

Pages that have been moved have aliases added to avoid 404s while google
search indexing catches up.