small update to Windows Error Reporting artifact (#1020)
Hi,

while using the artifact "Windows.System.WindowsErrorReporting" I
noticed a number of log entries regarding the results from the "glob()"
plugin - the field "FullPath" is to be replaced with the field "OSPath".

This small commit does the necessary changes for the artifact.
Tested the artifact on a Windows 11 system and saw no problems.

Kind Regards

Herbert