Create Windows.System.ModifyACL
This artifact modifies ACLs on files in the specified directory using the icacls command to grant full control to SYSTEM and deny access to Users and Everyone. This artifact can be very crucial given the case that velociraptor is deployed on a critical network/infrastructure, consequently, reducing cyber risk.

Since Velociraptor Agent Operates as SYSTEM, modifying its permissions to only allow SYSTEM user to access velociraptor will not cause any issues or conflicts.