Re-organize and update deployment docs (#1042)
The **Deployment section** is now oriented more towards guiding users
through the decision-making process and explaining the key concepts and
associated decisions rather than providing a step-by-step "cloud
deployment". The previous approach presented a potentially confusing
false dichotomy of "cloud deployment" vs. "self-signed SSL" which was
not very helpful to new users because most of the "cloud deployment"
instructions would also apply to on-premises deployments, and "cloud
deployments" could also use self-signed SSL (with specific caveats).

- More info added to Instant Velociraptor section as this is our
"quicker-than-quickstart" and people may prefer to try this first before
the Quickstart.

- The deployment section also now describes some unconventional ways of
using Velociraptor, which may provide strong alternatives to other DFIR
tools. Many potential users might be dismissive of Velociraptor if they
are not interested in the client-server deployment model and are not
aware that this is not the only way to use it.

**Quick Start Guide** - is intended to be more prescriptive and detailed
than the general deployment guide, as well as making fewer assumptions
about the user's skill level (therefore including steps that are not
obvious to Windows users such as `chmod +x`). It is intended to help new
users get a simple non-production deployment up and running as quickly
as possible. It deliberately avoids presenting choices as much as
possible, in order to keep the procedure simple. Discussion of such
choices/options is better suited to the Deployment Guide. It also
includes subtle steps to verify previous steps or ensure that the
further steps will work (i.e. catch issues in ways that are more
obvious). While the Quick Start is technically about deployment it is
not located in the Deployment section because a) users need to see it
immediately when they open the Documentation section, and b) the
Deployment docs are intended to contain explanations of concepts, not
How-to guides.

**"cloud deployment" transformed into a more general "deployment
example"**
- Since most of the "cloud deployment" article was spent describing
Google OAuth which is very specific and detailed, I've migrated that
into a separate KB article.
- Remove references to Google Domains and Google DDNS. We now support
CloudFlare and No-IP so these are mentioned instead.

**General**
- Update config wizard example to match current version.
- Provide instructions for both deb and rpm systems, i.e. remove Ubuntu
bias.
- Add links to the end of most articles to additionally guide users
(since there may be relevant info outside the menu structure) and ensure
that the relevant articles have a metadata summary to aid this approach.
- The process for generating client configs is now generally done from
the GUI rather than the CLI (e.g. the CLI wizard no longer produces a
client config), mainly to cater for orgs. So the docs now favour the GUI
as the primary method for obtaining the client config, although we also
describe the CLI method (including the orgs option).