Review and update Forensic Analysis section (#1074)
The main goal of this was to get rid of the obsolete url() path approach
when using the raw reg accessor. That section is completely reworked to
now use the modern remapping approach.

Reviewed and updated a few other pages in that section too. Pages with a
`last_reviewed` header are fully reviewed.

Other changes:

- Made sure that pages in the section have a summary, so that this can
be used to make the top level section page more navigable.
- Added more info to KB article about reconfiguring offline collectors,
especially the generic collector.
- Added some info to the CLI deadisk command page to explain the new
artifact-based mechanism.
- Added many more cross-reference links.
- Fixed up Matt's sloppy "Cloudflare Dynamic DNS" article :-)
- Added a note about the http_client() now natively supporting
multipart/form-data POST.
- Removed obsolete KB article about automating the creation of offline
collectors.
- Minor corrections and formatting improvements all over the place.