Export functions and reuse code in a new alert monitor
The original artifact has been completely rewritten with mostly the same features, except one can choose to get e-mails for failed flows even if they do not match filters.

A new artifact, Server.Monitoring.Alerts, sends e-mails for alerts sent with alert(). Just as the original artifact, it includes a bunch of client and flow context, in addition to the alert context. A parameter lets one extract a severity value from the alert context, since such a concept is not native to alert().