Add Linux.Network.Netstat.Watcher artifact (#1104)
Reading one of the latest analyses by Group-IB on the threat actor
UNC2891
[(https://www.group-ib.com/blog/unc2891-bank-heist/),](https://www.group-ib.com/blog/unc2891-bank-heist/)
I thought of creating an artifact that collects a one-time snapshot of
current non-LISTEN remote connections and a continuous diff stream of
added/removed/changed connections.

It is very useful for monitoring the connections that occur within a
time period set as a parameter.