Troubleshooting refactor (#1107)
This is primarily a refactor without adding much new content. But now
the gaps are more apparent!

The reasoning for consolidating this information under a dedicated
section is that troubleshooting spans many areas beyond deployment,
which is where most troubleshooting advice is currently located. Users
need to be able to get to at least a more specific category of
troubleshooting advice that's closer to their problem. The taxonomy I've
used here is closer, I think, to how I would categorize issues discussed
on the public forums. Some of these problem domains are not at all
deployment-related, for example troubleshooting VQL queries or hunts.

This is how it will eventually be, but I've hidden the pages/categories
that don't have enough content yet.

<img width="2992" height="2374" alt="docs_troubleshooting_"
src="https://github.com/user-attachments/assets/e1db4b07-80d3-4ef8-884e-1db0360385f7"
/>

All pages that were moved are also aliased so that links shared in the
past or search-indexed should not be broken by this change.