Collection container related updates (#1145)
- Added info on new (since 0.75.5) `decrypt` CLI command and update the
offline collector docs to make use of it.
- Added a new section to provide guidance about "importing vs. not
importing collections".
- Completed section about "Dead disk analysis on a collection container"
using the newly improved `Windows.Collectors.Remapping` artifact.
- Added some notes and cross-references for `fuse`, `unzip`, and
`decrypt` commands. Although the latest fuse changes are not yet in the
release builds, I doubt that it will cause any problems before they are
because very few people are using it.
- Added some extra info to KB article about backups.
- Removed some unnecesary "in recent versions" phrasing throughout the
docs.
- Migrated all the information from the KB article "How do I use the
files inside the offline collector ZIP?" into either Offline Collections
or the fuse container docs, and removed that KB article (and aliased
it's old path to the fuse article, which in turn links people to the
Offline Collections docs if they need that part of the information).
- Reviewed and revised most pages in the Searching Filesystems section.
- Moved the NTFS Analysis page to a more logical place in the content
hierarchy (and aliased it's old path).
- Updated CLI flags and associated docs for deb and rpm commands (the
output flag's purpose has changed).